+++ packages/zoetrope/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @supermousejs/zoetrope
 
+## 2.3.0
+
+### Patch Changes
+
+- Updated dependencies [f6f44b2]
+  - @supermousejs/core@2.2.0
+
 ## 2.2.0
 
 ### Patch Changes